# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "MultiTenant Api Documentation",
        "description": "MultiTenant Api Documentation",
        "contact": {
            "name": "Elianhers Blanco",
            "email": "elianhers.b@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api/v1"
        }
    ],
    "paths": {
        "/user": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Obtener listado de usuarios",
                "description": "listado de usuarios",
                "operationId": "c5892e9954ff5ccf5cac53a8cadb4650",
                "parameters": [
                    {
                        "name": "filter",
                        "in": "query",
                        "description": "filtrar usuario por nombre",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "default": "novo"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Numero de item por pagina",
                        "schema": {
                            "type": "integer",
                            "default": 10
                        }
                    },
                    {
                        "name": "direction",
                        "in": "query",
                        "description": "Sort direction (ASC or DESC)",
                        "schema": {
                            "type": "string",
                            "default": "ASC",
                            "enum": [
                                "ASC",
                                "DESC"
                            ]
                        }
                    },
                    {
                        "name": "sort_by",
                        "in": "query",
                        "description": "Sort field",
                        "schema": {
                            "type": "string",
                            "default": "name"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Respuesta exitosa",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/User"
                                            }
                                        },
                                        "meta": {
                                            "$ref": "#/components/schemas/Pagination"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Users"
                ],
                "summary": "Crear un nuevo usuario",
                "description": "Crear un nuevo usuario",
                "operationId": "create",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "description": "Nombre del usuario",
                                        "type": "string",
                                        "example": "novo"
                                    },
                                    "fqdn": {
                                        "description": "fqdn",
                                        "type": "string",
                                        "example": "novo"
                                    },
                                    "email": {
                                        "description": "Correo electrónico del usuario",
                                        "type": "string",
                                        "example": "novo@gmail.com"
                                    },
                                    "password": {
                                        "description": "Contraseña del usuario",
                                        "type": "string",
                                        "example": "123456"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Usuario creado exitosamente"
                    },
                    "422": {
                        "description": "Error de validación"
                    }
                }
            }
        },
        "/user/{id}": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Obtener usuario por ID",
                "operationId": "4cb26f403ff6b16406b58e88460726c7",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del usuario",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Usuario encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Usruario no encontrado"
                    }
                }
            },
            "put": {
                "tags": [
                    "Users"
                ],
                "summary": "Editar un nuevo usuario",
                "description": "Editar un nuevo usuario",
                "operationId": "update",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del usuario",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "description": "Nombre del usuario",
                                        "type": "string",
                                        "example": "novoTest"
                                    },
                                    "fqdn": {
                                        "description": "fqdn",
                                        "type": "string",
                                        "example": "novoTest"
                                    },
                                    "email": {
                                        "description": "Correo electrónico del usuario",
                                        "type": "string",
                                        "example": "novoTest@gmail.com"
                                    },
                                    "password": {
                                        "description": "Contraseña del usuario",
                                        "type": "string",
                                        "example": "1234567"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "202": {
                        "description": "Usuario Editado exitosamente"
                    },
                    "404": {
                        "description": "Usuario no encontrado"
                    },
                    "422": {
                        "description": "Error de validación"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Users"
                ],
                "summary": "Eliminar un usuario por ID",
                "description": "Eliminar un usuario por ID",
                "operationId": "8cf21a50c70c65e5ea560eab07171841",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del usuario a eliminar",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Usuario eliminado exitosamente"
                    },
                    "404": {
                        "description": "Usuario no encontrado"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Pagination": {
                "properties": {
                    "current_page": {
                        "type": "integer"
                    },
                    "last_page": {
                        "type": "integer"
                    },
                    "per_page": {
                        "type": "integer"
                    },
                    "total": {
                        "type": "integer"
                    }
                },
                "type": "object"
            },
            "GenericListCollection": {
                "properties": {
                    "data": {
                        "type": "array",
                        "items": {}
                    },
                    "meta": {
                        "$ref": "#/components/schemas/Pagination"
                    }
                },
                "type": "object"
            },
            "User": {
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "name": {
                        "type": "string"
                    },
                    "password": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string"
                    }
                },
                "type": "object"
            }
        }
    }
}